OPENEUROPA-1551: Fix small QA remarks.
